- Now we can upload multiple images to our product - When creating product, category cannot be None no longer - Now if upload folder does not exist , then it will be created - Product image deletion api replaced in "dashboard" blueprint - node_modules excluded from flake8
