+++ flaskshop/templates/products/_items.html
@@ -1,6 +1,6 @@
 {% for product in products %}
 <div class="col-6 col-lg-3 product-list">
-  <a href="{{ url_for("product.show", id=product.id) }}" class="link--clean">
+  <a href="{{ url_for('product.show', id=product.id) }}" class="link--clean">
     <div class="text-center">
       <div>
         <img class="img-responsive lazyload lazypreload" alt="" src="{{ product.first_img }}">
