+++ tox.ini
@@ -3,4 +3,4 @@
 max-line-length = 120
 ;exclude = tests/*
 ;max-complexity = 10
-exclude = .git,__pycache__,.venv/,migrations/,frontend/,docs/
+exclude = .git,__pycache__,.venv/,migrations/,frontend/,docs/,./node_modules/
